Update of Discover with Depth First Search attempt Also, attempt to make output more user readable Made seperate files for modularity and clean/orgainze clothes.
